Get oh-so tantalizingly close to being able to turn those Mono apps into native bundles This *ALMOST* works, but the problem is that upstream can't get their head on straight https://github.com/mono/mono/issues/10490
This issue has been open for FOUR YEARS and hasn't gotten any love or attention, and
it's the reason why we'll end up needing to use an older system with its own native
mono runtime bundled in. We can't cross-mkbuild here and it's awful.

But on the bright side, we do have a path forward